code improved, completed with this branch
