+++ .env.local
@@ -1,0 +1,17 @@
+# FSQ api keys
+# Project name: Coffee store api [FSQ]
+# name api key [coffee connoisseur app api key] value []
+NEXT_PUBLIC_FORSQUAREPLACES_APIKEY      = 
+NEXT_PUBLIC_FORSQUAREPLACES_URI         = ''
+
+# Environment
+NEXT_PUBLIC_NODE_ENV = "development" # [production, development, test]
+
+
+# unsplash api keys
+NEXT_PUBLIC_UNSPLASH_ACCESSKEY  = 
+
+
+NEXT_PUBLIC_INIT_LATLONG = ''
+# project name: First Test API key
+# API key (coffee store api key call testing):
